llama : control-vector -> adapter
