+++ src/llama-adapter.cpp
@@ -1,0 +1,5 @@
+#include "llama-adapter.h"
+
+void llama_lora_adapter_free(struct llama_lora_adapter * adapter) {
+    delete adapter;
+}
