--- src/llama-control-vector.cpp
@@ -1,1 +1,0 @@
-#include "llama-control-vector.h"
